+++ app/src/main/java/com/keeppieces/android/ui/bill/picker/AddCategoryDialog.kt
@@ -6,12 +6,10 @@ import android.content.Context
 import android.os.Bundle
 import android.view.Gravity
 import android.widget.TextView
-import android.widget.Toast
 import androidx.cardview.widget.CardView
 import androidx.fragment.app.DialogFragment
 import androidx.lifecycle.ViewModelProvider
 import com.keeppieces.android.R
-import com.keeppieces.android.logic.data.SecondaryCategory
 import com.keeppieces.android.ui.bill.BillViewModel
 
 //const val MaxCategoryLength = 4
@@ -126,8 +124,6 @@ class AddCategoryDialog(toPrimary: String, toSecondary: String) : DialogFragment
 //                    Toast.makeText(activity, "二级分类输入为空", Toast.LENGTH_LONG).show()
 //                }
             }
-            val secondaryLiveList = viewModel.findSecondaryList()
-
 
             builder.setTitle("添加类别")
                 .setView(view)
@@ -135,16 +131,7 @@ class AddCategoryDialog(toPrimary: String, toSecondary: String) : DialogFragment
                 ) { _, _ ->
 //                    addPrimary.text = textPrimary
 //                    addSecondary.text = textSecondary
-                    secondaryLiveList.observe(this, { tmp ->
-                        val secondaryList = tmp.map { t -> t.name }
-                        if (textSecondary in secondaryList) {
-                            Toast.makeText(activity, "已存在该二级分类", Toast.LENGTH_LONG).show()
-                            listener.onDialogPositiveClickForAddCategory(this)
-                        } else {
-                            viewModel.addSecondary(SecondaryCategory(textSecondary, textPrimary))
-                            listener.onDialogPositiveClickForAddCategory(this)
-                        }
-                    })
+                    listener.onDialogPositiveClickForAddCategory(this)
                 }
                 .setNegativeButton("取消"
                 ) { _, _ ->
